+++ src/Listeners/LogSession.php
@@ -36,15 +36,16 @@ class LogSession
      */
     public function handle(CallbackFormed $event)
     {
+        $uses = isset($event->response['controller']) ? $event->response['controller'] . '@' . $event->response['method'] : 'callback';
+
         $session = new Session();
         $session->user_id = $event->user->id;
         $session->update_id = $event->response['update_id'];
         $session->event = $event->response['event'];
         $session->listener = $event->response['listener'];
-        $session->hook = $event->response['hook'];
-        $session->controller = $event->response['controller'];
-        $session->method = $event->response['method'];
-        $session->last_state = $event->response['state'];
+        $session->contains = $event->response['contains'];
+        $session->uses = $uses;
+        $session->location = $event->response['state'];
         $session->save();
     }
 }