+++ src/Support/FormRequest.php
@@ -27,10 +27,10 @@ class FormRequest
      *
      * @param array $request
      * @param array $route
-     * @param string|null $state
+     * @param string|null $location
      * @return \Wekser\Laragram\BotRequest
      */
-    public function setRequest(array $request, array $route, ?string $state = null): BotRequest
+    public function setRequest(array $request, array $route, ?string $location = null): BotRequest
     {
         $type = collect($request)->search(function ($value, $key) {
             return is_array($value) && isset($value['from']);
@@ -39,13 +39,12 @@ class FormRequest
         $entity = $request[$type];
 
         $this->request['all'] = $entity;
-        $this->request['controller'] = $route['controller'];
         $this->request['event'] = $type;
-        $this->request['query'] = collect($entity)->get($route['listener']);
-        $this->request['hook'] = $route['hook'] ?? null;
         $this->request['listener'] = $route['listener'];
-        $this->request['method'] = $route['method'];
-        $this->request['state'] = $route['alias'] ?? $state;
+        $this->request['input'] = collect($entity)->get($route['listener']);
+        $this->request['contains'] = $route['contains'] ?? null;
+        $this->request['uses'] = isset($route['controller']) ? $route['controller'] . '@' . $route['method'] : 'callback';
+        $this->request['location'] = $route['from'] ?? $location;
         $this->request['update_id'] = $request['update_id'];
 
         return new BotRequest($this->request);