+++ CHANGELOG.md
@@ -5,11 +5,23 @@ All notable changes to `Laragram` will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## [v1.3.1] (2022-11-17)
+## [v1.4.0] (2022-11-18)
 
 ### Added
 
 - Added customization User model in `config.php`
+- Added the ability to run callback functions in routes
+- Added format settings in `BotResponse::text()`
+- Added the ability to «basic-text» return
+
+### Changed
+
+- Changed router method names:
+    - bind -> get
+    - catch -> contains
+    - alias -> from
+- Changed database table `laragram_sessions`
+- Improved code
 
 ## [v1.3.0] (2022-11-12)